Move FieldError into validation folder [common]
